+++ src/main/kotlin/omok/model/adapter/RuleAdapter.kt
@@ -1,6 +1,7 @@
 package omok.model.adapter
 
-import omok.model.Stone
+import omok.model.game.FoulCondition
+import omok.model.stone.Stone
 
 interface RuleAdapter {
     fun checkAnyFoulCondition(
