+++ src/main/kotlin/omok/model/stone/Point.kt
@@ -1,3 +1,3 @@
-package omok.model
+package omok.model.stone
 
 data class Point(val row: Int, val col: Int)
