+++ src/main/kotlin/omok/model/stone/Stone.kt
@@ -1,4 +1,4 @@
-package omok.model
+package omok.model.stone
 
 data class Stone(
     val point: Point,
